rename reporter package to bitcoinspv
