+++ bitcoinspv/config/common.go
