+++ bitcoinspv/expected_bitcoin_client.go
@@ -1,4 +1,4 @@
-package reporter
+package bitcoinspv
 
 import (
 	"github.com/btcsuite/btcd/btcjson"
@@ -6,7 +6,7 @@ import (
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/btcsuite/btcd/wire"
 
-	"github.com/gonative-cc/relayer/reporter/types"
+	"github.com/gonative-cc/relayer/bitcoinspv/types"
 )
 
 type BTCClient interface {
