+++ bitcoinspv/expected_native_client.go
@@ -1,9 +1,9 @@
-package reporter
+package bitcoinspv
 
 import (
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/btcsuite/btcd/wire"
-	"github.com/gonative-cc/relayer/reporter/types"
+	"github.com/gonative-cc/relayer/bitcoinspv/types"
 )
 
 // NOTE: not copied
