+++ bitcoinspv/reporter_metrics.go
@@ -1,4 +1,4 @@
-package reporter
+package bitcoinspv
 
 import (
 	"time"
