+++ bitcoinspv/retry_test.go
@@ -1,4 +1,4 @@
-package reporter
+package bitcoinspv
 
 import (
 	"testing"
