+++ bitcoinspv/types/blockevent.go
