+++ bitcoinspv/types/btc_header_bytes.go
