+++ bitcoinspv/types/btc_headers_hash_bytes.go
