+++ bitcoinspv/types/btc_utils.go
