+++ bitcoinspv/types/btccache.go
