+++ bitcoinspv/types/btclightclient.go
