+++ bitcoinspv/types/config.go
