+++ bitcoinspv/types/errors.go
