+++ bitcoinspv/types/spv_proof.go
