+++ bitcoinspv/types/utils.go
